Разделили презентер фильмов на два отдельных
